Restores package structure; makes unredact.py a wrapper.
