+++ unredact/__init__.py
@@ -1,0 +1,1 @@
+from unredact.__main__ import get_output_filename, main
